src/transcription/requirements.txt
# Production requirements for MedTranscribe
streamlit>=1.28.0
openai-whisper>=20231117
openai>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
packaging>=23.0
typing-extensions>=4.5.0 